+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -13,21 +13,12 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.databinding.DataBindingUtil
 import woowacourse.shopping.R
-import woowacourse.shopping.data.cart.CartDatabase
-import woowacourse.shopping.data.cart.CartRepositoryImpl
 import woowacourse.shopping.databinding.ActivityCartBinding
-import woowacourse.shopping.ui.viewmodel.CartViewModel
-import woowacourse.shopping.utils.ViewModelFactory
 
 class CartActivity : AppCompatActivity() {
     private lateinit var binding: ActivityCartBinding
 
-    private val viewModel: CartViewModel by viewModels {
-        ViewModelFactory.createCartViewModelFactory(
-            CartRepositoryImpl(CartDatabase.getInstance(this)),
-        )
-    }
-
+    private val viewModel: CartViewModel by viewModels { CartViewModel.Factory }
     private val cartAdapter: CartAdapter by lazy { initAdapter() }
 
     override fun onCreate(savedInstanceState: Bundle?) {
